# docs/swagger.json
{"openapi":"3.0.0","info":{"title":"Tourist Guides API","version":"1.0.0","description":"API RESTful para gestión de usuarios, guías, reservas y servicios relacionados."},"servers":[{"url":"http://localhost:4000","description":"Local dev"}],"paths":{"/api":{"get":{"summary":"Info básica de la API","responses":{"200":{"description":"OK"}}}},"/api/health":{"get":{"summary":"Health check","responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"summary":"Login","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLogin"}}}},"responses":{"200":{"description":"Tokens y usuario"}}}},"/api/auth/register":{"post":{"summary":"Register","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegister"}}}},"responses":{"201":{"description":"Usuario creado"}}}},"/api/auth/refresh":{"post":{"summary":"Refresh token","responses":{"200":{"description":"Nuevo token"}}}},"/api/auth/profile":{"get":{"summary":"Perfil (requires auth)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Perfil"}}},"put":{"summary":"Actualizar perfil (requires auth)","security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Perfil actualizado"}}}},"/api/users":{"get":{"summary":"Listar usuarios","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/users/{id}":{"get":{"summary":"Usuario por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Usuario"}}}},"/api/guides":{"get":{"summary":"Listar guías","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/guides/{id}":{"get":{"summary":"Guía por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Guía"}}}},"/api/guides/{id}/specialties":{"get":{"summary":"Especialidades de un guía","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/guides/{id}/languages":{"get":{"summary":"Idiomas de un guía","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/bookings":{"get":{"summary":"Listar reservas","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/bookings/my":{"get":{"summary":"Mis reservas","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/bookings/{id}":{"get":{"summary":"Reserva por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Reserva"}}}},"/api/zones":{"get":{"summary":"Listar zonas","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/zones/{id}":{"get":{"summary":"Zona por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Zona"}}}},"/api/tourist-places":{"get":{"summary":"Listar lugares turísticos","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/tourist-places/{id}":{"get":{"summary":"Lugar por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Lugar"}}}},"/api/specialties":{"get":{"summary":"Listar especialidades","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/languages":{"get":{"summary":"Listar idiomas","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/payments":{"get":{"summary":"Listar pagos","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/payments/{id}":{"get":{"summary":"Pago por id","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Pago"}}}},"/api/favorites":{"get":{"summary":"Listar favoritos","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/favorites/guides":{"get":{"summary":"Favoritos - guías","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/favorites/zones":{"get":{"summary":"Favoritos - zonas","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/notifications":{"get":{"summary":"Listar notificaciones","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/weather-alerts":{"get":{"summary":"Listar alertas clima","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lista"}}}},"/api/hospedajes":{"get":{"summary":"Listar hospedajes","responses":{"200":{"description":"Lista de hospedajes"}}},"post":{"summary":"Crear hospedaje","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HospedajeInput"}}}},"responses":{"201":{"description":"Hospedaje creado"}}}},"/api/hospedajes/{hospedajeId}":{"get":{"summary":"Obtener hospedaje por id","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Hospedaje"}}},"put":{"summary":"Actualizar hospedaje","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HospedajeInput"}}}},"responses":{"200":{"description":"Actualizado"}}},"delete":{"summary":"Eliminar hospedaje","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"description":"Eliminado"}}}},"/api/hospedajes/{hospedajeId}/alimentacion":{"get":{"summary":"Listar opciones de alimentación para un hospedaje","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Lista de opciones"}}},"post":{"summary":"Crear opción de alimentación para un hospedaje","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlimentacionInput"}}}},"responses":{"201":{"description":"Opción creada"}}}},"/api/hospedajes/{hospedajeId}/alimentacion/{id}":{"get":{"summary":"Obtener opción de alimentación por id","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Opción"}}},"put":{"summary":"Actualizar opción de alimentación","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlimentacionInput"}}}},"responses":{"200":{"description":"Actualizado"}}},"delete":{"summary":"Eliminar opción de alimentación","parameters":[{"name":"hospedajeId","in":"path","required":true,"schema":{"type":"integer"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"description":"Eliminado"}}}},"/api/home/stats":{"get":{"summary":"Estadísticas para home","responses":{"200":{"description":"Home stats"}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"AuthLogin":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthRegister":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password"]},"UserUpdate":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"}}},"HospedajeInput":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"zone_id":{"type":"integer"},"guide_id":{"type":"integer"},"description":{"type":"string"},"price":{"type":"number"}},"required":["name","address"]},"AlimentacionInput":{"type":"object","properties":{"type":{"type":"string"},"price":{"type":"number"},"notes":{"type":"string"}},"required":["type","price"]}}}}